pdf_parser/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.0

# PDF Processing (Core functionality)
pdfplumber==0.10.3
pymupdf==1.23.26
pillow==10.1.0
pdf2image==1.16.3

# LLM Integration (AI functionality)
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.8.3

# OCR Support (Railway compatible)
pytesseract==0.3.10
opencv-python-headless==4.9.0.80

# Data processing
pandas==2.1.4
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0

# Optional (remove if causing issues)
# python-magic==0.4.27
# redis==5.0.1
# sqlalchemy==2.0.27
# alembic==1.13.1
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
